Crate layout_thread

The layout thread. Performs layout on the DOM, builds display lists and sends them to be painted.

Modules

dom_wrapper
A safe wrapper for DOM nodes that prevents layout from mutating the DOM, from letting DOM nodes escape, and from generally doing anything that it isn't supposed to. This is accomplished via a simple whitelist of allowed operations, along with some lifetime magic to prevent nodes from escaping.

Structs

LayoutThread
Information needed by the layout thread.

RegisteredPainterImpl

RegisteredPaintersImpl

RwData

ScriptReflowResult

UA_STYLESHEETS

Enums

RWGuard
The LayoutThread rw_data lock must remain locked until the first reflow, as RPC calls don't make sense until then. Use this in combination with LayoutThread::lock_rw_data and LayoutThread::return_rw_data.

Functions

add_font_face_rules

get_root_flow_background_color

get_ua_stylesheets