Struct layout_thread::dom_wrapper::ServoLayoutNode

pub struct ServoLayoutNode<'a> {
    node: LayoutDom<Node>,
    chain: PhantomData<&'a ()>,
}

Fields
node: LayoutDom<Node>

The wrapped node.

chain: PhantomData<&'a ()>

Being chained to a PhantomData prevents LayoutNodes from escaping.

Methods

impl<'ln> ServoLayoutNode<'ln>

fn from_layout_js(n: LayoutDom<Node>) -> ServoLayoutNode<'ln>

pub unsafe fn new(address: &TrustedNodeAddress) -> ServoLayoutNode

pub unsafe fn new_with_this_lifetime(
&self,
node: &LayoutDom<Node>
) -> ServoLayoutNode<'ln>

Creates a new layout node with the same lifetime as this layout node.

fn script_type_id(&self) -> NodeTypeId

impl<'ln> ServoLayoutNode<'ln>

pub unsafe fn get_jsmanaged(&self) -> &LayoutDom<Node>

Returns the interior of this node as a LayoutDom. This is highly unsafe for layout to call and as such is marked unsafe.