Struct layout_thread::dom_wrapper::ServoThreadSafeLayoutNode

pub struct ServoThreadSafeLayoutNode<'ln> {
    node: ServoLayoutNode<'ln>,
    pseudo: PseudoElementType,
}

Fields
node: ServoLayoutNode<'ln>

The wrapped node.

pseudo: PseudoElementType

The pseudo-element type, with (optionally) a specified display value to override the stylesheet.

Methods

impl<'ln> ServoThreadSafeLayoutNode<'ln>

pub unsafe fn new_with_this_lifetime(
&self,
node: &LayoutDom<Node>
) -> ServoThreadSafeLayoutNode<'ln>

Creates a new layout node with the same lifetime as this layout node.

pub fn new<'a>(node: &ServoLayoutNode<'a>) -> ServoThreadSafeLayoutNode<'a>

Creates a new ServoThreadSafeLayoutNode from the given ServoLayoutNode.

unsafe fn get_jsmanaged(&self) -> &LayoutDom<Node>

Returns the interior of this node as a LayoutDom. This is highly unsafe for layout to call and as such is marked unsafe.