Function layout_thread::reflow_query_type_needs_display_list

fn reflow_query_type_needs_display_list(query_type: &ReflowQueryType) -> bool

Returns true if the given reflow query type needs a full, up-to-date display list to be present or false if it only needs stacking-relative positions.