Struct layout_thread::LayoutThread

pub struct LayoutThread {
    id: PipelineId,
    url: ServoUrl,
    is_iframe: bool,
    port: Receiver<Msg>,
    pipeline_port: Receiver<LayoutControlMsg>,
    font_cache_receiver: Receiver<()>,
    font_cache_sender: IpcSender<()>,
    constellation_chan: IpcSender<ConstellationMsg>,
    script_chan: IpcSender<ConstellationControlMsg>,
    time_profiler_chan: ProfilerChan,
    mem_profiler_chan: ProfilerChan,
    image_cache_thread: ImageCacheThread,
    font_cache_thread: FontCacheThread,
    first_reflow: bool,
    parallel_traversal: Option<ThreadPool>,
    parallel_flag: bool,
    generation: u32,
    new_animations_sender: Sender<Animation>,
    new_animations_receiver: Receiver<Animation>,
    outstanding_web_fonts: Arc<AtomicUsize>,
    root_flow: Option<FlowRef>,
    running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    epoch: Epoch,
    viewport_size: Size2D<Au>,
    rw_data: Arc<Mutex<LayoutThreadData>>,
    error_reporter: CSSErrorReporter,
    webrender_image_cache: Arc<RwLock<HashMap<(ServoUrl, UsePlaceholder), WebRenderImageInfo, BuildHasherDefault<FnvHasher>>>>,
    webrender_api: RenderApi,
    timer: Timer,
    layout_threads: usize,
    quirks_mode: Option<QuirksMode>,
}

Information needed by the layout thread.

Fields
id: PipelineId

The ID of the pipeline that we belong to.

url: ServoUrl

The URL of the pipeline that we belong to.

is_iframe: bool

Is the current reflow of an iframe, as opposed to a root window?

port: Receiver<Msg>

The port on which we receive messages from the script thread.

pipeline_port: Receiver<LayoutControlMsg>

The port on which we receive messages from the constellation.

font_cache_receiver: Receiver<()>

The port on which we receive messages from the font cache thread.

font_cache_sender: IpcSender<()>

The channel on which the font cache can send messages to us.

constellation_chan: IpcSender<ConstellationMsg>

The channel on which messages can be sent to the constellation.

script_chan: IpcSender<ConstellationControlMsg>

The channel on which messages can be sent to the script thread.

time_profiler_chan: ProfilerChan

The channel on which messages can be sent to the time profiler.

mem_profiler_chan: ProfilerChan

The channel on which messages can be sent to the memory profiler.

image_cache_thread: ImageCacheThread

The channel on which messages can be sent to the image cache.

font_cache_thread: FontCacheThread

Public interface to the font cache thread.

first_reflow: bool

Is this the first reflow in this LayoutThread?

parallel_traversal: Option<ThreadPool>

The workers that we use for parallel operation.

parallel_flag: bool

Flag to indicate whether to use parallel operations

generation: u32

Starts at zero, and increased by one every time a layout completes. This can be used to easily check for invalid stale data.

new_animations_sender: Sender<Animation>

A channel on which new animations that have been triggered by style recalculation can be sent.

new_animations_receiver: Receiver<Animation>

Receives newly-discovered animations.

outstanding_web_fonts: Arc<AtomicUsize>

The number of Web fonts that have been requested but not yet loaded.

root_flow: Option<FlowRef>

The root of the flow tree.

running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

The list of currently-running animations.

expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>

The list of animations that have expired since the last style recalculation.

epoch: Epoch

A counter for epoch messages

viewport_size: Size2D<Au>

The size of the viewport. This may be different from the size of the screen due to viewport constraints.

rw_data: Arc<Mutex<LayoutThreadData>>

A mutex to allow for fast, read-only RPC of layout's internal data structures, while still letting the LayoutThread modify them.

All the other elements of this struct are read-only.

error_reporter: CSSErrorReporter

The CSS error reporter for all CSS loaded in this layout thread

webrender_image_cache: Arc<RwLock<HashMap<(ServoUrl, UsePlaceholder), WebRenderImageInfo, BuildHasherDefault<FnvHasher>>>> webrender_api: RenderApi timer: Timer

The timer object to control the timing of the animations. This should only be a test-mode timer during testing for animations.

layout_threads: usize quirks_mode: Option<QuirksMode>

Which quirks mode are we rendering the document in?

Methods

impl LayoutThread

fn new(id: PipelineId,
url: ServoUrl,
is_iframe: bool,
port: Receiver<Msg>,
pipeline_port: IpcReceiver<LayoutControlMsg>,
constellation_chan: IpcSender<ConstellationMsg>,
script_chan: IpcSender<ConstellationControlMsg>,
image_cache_thread: ImageCacheThread,
font_cache_thread: FontCacheThread,
time_profiler_chan: ProfilerChan,
mem_profiler_chan: ProfilerChan,
webrender_api_sender: RenderApiSender,
layout_threads: usize)
-> LayoutThread

Creates a new LayoutThread structure.

fn start(self)

Starts listening on the port.

fn build_layout_context(&self,
rw_data: &LayoutThreadData,
request_images: bool)
-> LayoutContext

fn handle_request<'a, 'b>(&mut self,
possibly_locked_rw_data: &mut RwData<'a, 'b>)
-> bool

Receives and dispatches messages from the script and constellation threads

fn handle_request_helper<'a, 'b>(&mut self,
request: Msg,
possibly_locked_rw_data: &mut RwData<'a, 'b>)
-> bool

Receives and dispatches messages from other threads.

fn collect_reports<'a, 'b>(&self,
reports_chan: ReportsChan,
possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn create_layout_thread(&self, info: NewLayoutThreadInfo)

fn prepare_to_exit(&mut self, response_chan: Sender<()>)

Enters a quiescent state in which no new messages will be processed until an ExitNow is received. A pong is immediately sent on the given response channel.

fn exit_now(&mut self)

Shuts down the layout thread now. If there are any DOM nodes left, layout will now (safely) crash.

fn handle_add_stylesheet<'a, 'b>(&self,
stylesheet: Arc<Stylesheet>,
possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn handle_advance_clock_ms<'a, 'b>(&mut self,
how_many_ms: i32,
possibly_locked_rw_data: &mut RwData<'a, 'b>,
tick_animations: bool)

Advances the animation clock of the document.

fn handle_set_quirks_mode<'a, 'b>(&self,
possibly_locked_rw_data: &mut RwData<'a, 'b>)

Sets quirks mode for the document, causing the quirks mode stylesheet to be used.

fn try_get_layout_root<N: LayoutNode>(&self, node: N) -> Option<FlowRef>

fn solve_constraints(layout_root: &mut Flow, layout_context: &LayoutContext)

Performs layout constraint solving.

This corresponds to Reflow() in Gecko and layout() in WebKit/Blink and should be benchmarked against those two. It is marked #[inline(never)] to aid profiling.

fn solve_constraints_parallel(traversal: &ThreadPool,
layout_root: &mut Flow,
profiler_metadata: Option<TimerMetadata>,
time_profiler_chan: ProfilerChan,
layout_context: &LayoutContext)

Performs layout constraint solving in parallel.

This corresponds to Reflow() in Gecko and layout() in WebKit/Blink and should be benchmarked against those two. It is marked #[inline(never)] to aid profiling.

fn compute_abs_pos_and_build_display_list(&mut self,
data: &Reflow,
query_type: Option<&ReflowQueryType>,
document: Option<&ServoLayoutDocument>,
layout_root: &mut Flow,
layout_context: &mut LayoutContext,
rw_data: &mut LayoutThreadData)

Computes the stacking-relative positions of all flows and, if the painting is dirty and the reflow goal and query type need it, builds the display list.

fn handle_reflow<'a, 'b>(&mut self,
data: &ScriptReflow,
possibly_locked_rw_data: &mut RwData<'a, 'b>)

The high-level routine that performs layout threads.

fn respond_to_query_if_necessary(&mut self,
query_type: &ReflowQueryType,
rw_data: &mut LayoutThreadData,
context: &mut LayoutContext)

fn set_stacking_context_scroll_states<'a, 'b>(&mut self,
new_scroll_states: Vec<StackingContextScrollState>,
possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn tick_all_animations<'a, 'b>(&mut self,
possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn tick_animations(&mut self, rw_data: &mut LayoutThreadData)

fn reflow_with_newly_loaded_web_font<'a, 'b>(&mut self,
possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn perform_post_style_recalc_layout_passes(&mut self,
data: &Reflow,
query_type: Option<&ReflowQueryType>,
document: Option<&ServoLayoutDocument>,
rw_data: &mut LayoutThreadData,
context: &mut LayoutContext)

fn perform_post_main_layout_passes(&mut self,
data: &Reflow,
query_type: Option<&ReflowQueryType>,
document: Option<&ServoLayoutDocument>,
rw_data: &mut LayoutThreadData,
layout_context: &mut LayoutContext)

fn reflow_all_nodes(flow: &mut Flow)

fn profiler_metadata(&self) -> Option<TimerMetadata>

Returns profiling information which is passed to the time profiler.