Struct layout_thread::dom_wrapper::ServoThreadSafeLayoutElement

pub struct ServoThreadSafeLayoutElement<'le> {
    element: ServoLayoutElement<'le>,
    pseudo: PseudoElementType,
}

A wrapper around elements that ensures layout can only ever access safe properties and cannot race on elements.

Fields
element: ServoLayoutElement<'le>pseudo: PseudoElementType

The pseudo-element type, with (optionally) a specified display value to override the stylesheet.