Struct layout_thread::dom_wrapper::ServoThreadSafeLayoutNode

pub struct ServoThreadSafeLayoutNode<'ln> {
    node: ServoLayoutNode<'ln>,
    pseudo: PseudoElementType,
}

Fields
node: ServoLayoutNode<'ln>

The wrapped node.

pseudo: PseudoElementType

The pseudo-element type, with (optionally) a specified display value to override the stylesheet.

Implementations

impl<'ln> ServoThreadSafeLayoutNode<'ln>

pub fn new(node: ServoLayoutNode<'ln>) -> Self

Creates a new ServoThreadSafeLayoutNode from the given ServoLayoutNode.

unsafe fn get_jsmanaged(&self) -> LayoutDom<'ln, Node>

Returns the interior of this node as a LayoutDom. This is highly unsafe for layout to call and as such is marked unsafe.